Hospitals to be registered/metro-medical/requirements.txt
Django==4.2.10
PyJWT==2.8.0
